Make tests run with their own server as a fixture
This is all pretty complicated, but there's no getting around somehow
waiting for the server to be ready for connections. It could also be
done by starting the server separately and writing some shell code to
wait for 'Server is listening', but being able to start the server as
part of running the tests is also convenient locally.

Note that this now starts a new server for every test, which will
obviously need to change later.